Fix RUBY_FFI-38 - was a cut-n-paste typo, long as aliased to int32 instead of int64

+++ ext/ffi_c/AbstractMemory.c
@@ -373,7 +373,7 @@ rbffi_AbstractMemory_Init(VALUE moduleFFI)
     if (sizeof(long) == 4) {
         ALIAS(long, int32);
     } else {
-        ALIAS(long, int32);
+        ALIAS(long, int64);
     }
 
     rb_define_method(classMemory, "put_float32", memory_put_float32, 2);
